requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-countries==3.3
django-ipware==1.1.1
django-pylibmc==0.6.0
django-storages==1.1.8
django-toolbelt==0.0.1
djangoappengine==1.7.1
djangorestframework==3.2.4
djangotoolbox==1.8.0
geojson==1.3.1
google-api-python-client==1.4.2
gunicorn==19.3.0
httplib2==0.9.2
MySQL-python==1.2.5
newrelic==2.56.0.42
oauth2client==1.5.1
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pylibmc==1.5.0
PyMySQL==0.6.7
requests==2.7.0
rsa==3.2
simplejson==3.8.0
six==1.10.0
static3==0.6.1
uritemplate==0.6
whitenoise==2.0.4